requirements.txt
fastapi>=0.100.0
pydantic>=2.0.0
uvicorn>=0.15.0
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
python-dotenv>=1.0.1
alembic>=1.7.0
redis>=4.0.0
pytest==8.2.2
httpx>=0.18.0
emails>=0.6.0
psycopg2-binary>=2.9.0
email-validator>=2.0.0
resend==0.7.0
greenlet>=3.0.0
freezegun 